requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-CORS>=3.0.10
Flask-Login>=0.6.2
python-dotenv>=1.0.0
Werkzeug>=2.3.7
transformers>=4.30.0
scikit-learn>=1.0.0
xgboost>=1.5.0
numpy>=1.21.0
ollama>=0.1.6
pandas>=1.3.0
xlsxwriter>=3.0.0
tensorflow>=2.9.0
prophet>=1.1.0
statsmodels>=0.13.0
